fixed warnings

+++ lootgame/lootgame/game.cpp
@@ -739,8 +739,8 @@ static void _cameraFollowPlayer(GameState& g) {
       //camera follow on dude
       vp.x = clamp(dudePos.x - (vp.w / 2), 0, g.map->size.x - vp.w);
       vp.y = clamp(dudePos.y - (vp.h / 2), 0, g.map->size.y - vp.h);
-      vp.w = Const.vpSize.x;
-      vp.h = Const.vpSize.y;
+      vp.w = (float)Const.vpSize.x;
+      vp.h = (float)Const.vpSize.y;
    }
 }
